refactor: split profiler modules
