+++ crates/ox_content_profiler/src/alloc.rs
@@ -13,6 +13,18 @@
 use std::alloc::{GlobalAlloc, Layout, System};
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 
+mod counter;
+mod histogram;
+mod snapshot;
+
+#[cfg(test)]
+mod tests;
+
+pub use counter::AllocCounter;
+use histogram::size_class_bucket;
+pub use histogram::SizeHistogram;
+pub use snapshot::{AllocDelta, AllocSnapshot};
+
 const SIZE_CLASS_BUCKETS: usize = 32;
 
 #[allow(unsafe_code)]
@@ -225,200 +237,5 @@ fn record_dealloc(size: u64) {
             Ok(_) => break,
             Err(observed) => cur = observed,
         }
-    }
-}
-
-#[inline]
-fn size_class_bucket(size: u64) -> usize {
-    // Bucket by floor(log2(size)). Bucket 0 catches size 0 (rare but legal),
-    // bucket 1 catches sizes 1..=1, bucket k catches sizes [2^(k-1), 2^k).
-    if size == 0 {
-        return 0;
-    }
-    let log2 = (u64::BITS - 1 - size.leading_zeros()) as usize;
-    log2.min(SIZE_CLASS_BUCKETS - 1) + usize::from(size > 1)
-}
-
-/// Instantaneous snapshot of allocator counters.
-#[derive(Debug, Clone)]
-pub struct AllocSnapshot {
-    pub allocations: u64,
-    pub deallocations: u64,
-    pub bytes_allocated: u64,
-    pub bytes_deallocated: u64,
-    pub current_live_bytes: u64,
-    pub peak_live_bytes: u64,
-    pub largest_single_alloc: u64,
-    pub size_class_buckets: [u64; SIZE_CLASS_BUCKETS],
-}
-
-impl AllocSnapshot {
-    #[must_use]
-    pub fn capture() -> Self {
-        let mut buckets = [0u64; SIZE_CLASS_BUCKETS];
-        for (i, slot) in STATS.size_class_buckets.iter().enumerate() {
-            buckets[i] = slot.load(Ordering::Relaxed);
-        }
-        Self {
-            allocations: STATS.allocations.load(Ordering::Relaxed),
-            deallocations: STATS.deallocations.load(Ordering::Relaxed),
-            bytes_allocated: STATS.bytes_allocated.load(Ordering::Relaxed),
-            bytes_deallocated: STATS.bytes_deallocated.load(Ordering::Relaxed),
-            current_live_bytes: STATS.current_live_bytes.load(Ordering::Relaxed),
-            peak_live_bytes: STATS.peak_live_bytes.load(Ordering::Relaxed),
-            largest_single_alloc: STATS.largest_single_alloc.load(Ordering::Relaxed),
-            size_class_buckets: buckets,
-        }
-    }
-
-    /// Compute the change between `before` and `self` (`self` is the later
-    /// observation).
-    #[must_use]
-    pub fn delta_from(&self, before: &Self) -> AllocDelta {
-        let mut buckets = [0u64; SIZE_CLASS_BUCKETS];
-        for (i, slot) in buckets.iter_mut().enumerate() {
-            *slot = self.size_class_buckets[i].saturating_sub(before.size_class_buckets[i]);
-        }
-        AllocDelta {
-            allocations: self.allocations.saturating_sub(before.allocations),
-            deallocations: self.deallocations.saturating_sub(before.deallocations),
-            bytes_allocated: self.bytes_allocated.saturating_sub(before.bytes_allocated),
-            bytes_deallocated: self.bytes_deallocated.saturating_sub(before.bytes_deallocated),
-            // Peak above baseline: how much higher did the peak go vs. the
-            // peak at snapshot time? This is the closest single-counter
-            // approximation to "max additional live bytes during the scope"
-            // without resetting shared state.
-            peak_above_baseline: self.peak_live_bytes.saturating_sub(before.peak_live_bytes),
-            ending_live_bytes: self.current_live_bytes,
-            starting_live_bytes: before.current_live_bytes,
-            largest_single_alloc: self.largest_single_alloc.max(before.largest_single_alloc),
-            size_class_buckets: SizeHistogram { buckets },
-        }
-    }
-}
-
-/// Difference between two [`AllocSnapshot`]s.
-#[derive(Debug, Clone)]
-pub struct AllocDelta {
-    pub allocations: u64,
-    pub deallocations: u64,
-    pub bytes_allocated: u64,
-    pub bytes_deallocated: u64,
-    /// Max additional live bytes observed during the window.
-    pub peak_above_baseline: u64,
-    pub starting_live_bytes: u64,
-    pub ending_live_bytes: u64,
-    pub largest_single_alloc: u64,
-    pub size_class_buckets: SizeHistogram,
-}
-
-impl AllocDelta {
-    /// Net change in live bytes (ending - starting). Can be negative-looking
-    /// when more was freed than allocated; returns saturating signed-ish u64
-    /// for simplicity. Use [`AllocDelta::net_growth`] for clarity.
-    #[must_use]
-    pub fn net_growth(&self) -> i128 {
-        i128::from(self.ending_live_bytes) - i128::from(self.starting_live_bytes)
-    }
-}
-
-/// Histogram of allocations bucketed by power-of-two size class.
-#[derive(Debug, Clone)]
-pub struct SizeHistogram {
-    pub buckets: [u64; SIZE_CLASS_BUCKETS],
-}
-
-impl SizeHistogram {
-    /// Returns the human-readable label for bucket `i`, e.g. `"32..64"`.
-    #[must_use]
-    pub fn bucket_label(i: usize) -> String {
-        if i == 0 {
-            return "0".into();
-        }
-        if i == 1 {
-            return "1".into();
-        }
-        let lo = 1u64 << (i - 1);
-        let hi = 1u64 << i.min(63);
-        let mut label = String::with_capacity(42);
-        push_u64(&mut label, lo);
-        label.push_str("..");
-        push_u64(&mut label, hi);
-        label
-    }
-
-    /// Iterates `(label, count)` pairs, skipping empty buckets.
-    pub fn iter_nonempty(&self) -> impl Iterator<Item = (String, u64)> + '_ {
-        self.buckets
-            .iter()
-            .enumerate()
-            .filter(|(_, c)| **c > 0)
-            .map(|(i, c)| (Self::bucket_label(i), *c))
-    }
-}
-
-fn push_u64(output: &mut String, value: u64) {
-    let mut buffer = [0_u8; 20];
-    let mut cursor = buffer.len();
-    let mut rest = value;
-
-    loop {
-        cursor -= 1;
-        buffer[cursor] = b'0' + (rest % 10) as u8;
-        rest /= 10;
-        if rest == 0 {
-            break;
-        }
-    }
-
-    let digits = std::str::from_utf8(&buffer[cursor..]).expect("digits are valid utf-8");
-    output.push_str(digits);
-}
-
-/// Convenience guard: takes a baseline snapshot on construction, exposes the
-/// running delta until dropped. Useful for ad-hoc scope measurements outside
-/// of [`crate::Recorder`].
-pub struct AllocCounter {
-    baseline: AllocSnapshot,
-}
-
-impl AllocCounter {
-    #[must_use]
-    pub fn start() -> Self {
-        Self { baseline: AllocSnapshot::capture() }
-    }
-
-    #[must_use]
-    pub fn delta(&self) -> AllocDelta {
-        AllocSnapshot::capture().delta_from(&self.baseline)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // We can't safely install ourselves as the global allocator in these
-    // tests (the cfg is per-binary), but we can still exercise the snapshot
-    // arithmetic and the bucket layout.
-
-    #[test]
-    fn bucket_layout_covers_typical_sizes() {
-        let mut hist = SizeHistogram { buckets: [0; SIZE_CLASS_BUCKETS] };
-        for &size in &[0u64, 1, 2, 7, 16, 17, 64, 1024, 4096, 65_536] {
-            let b = size_class_bucket(size);
-            hist.buckets[b] += 1;
-        }
-        let count = hist.iter_nonempty().count();
-        assert!(count > 0);
-    }
-
-    #[test]
-    fn delta_handles_no_change() {
-        let snap = AllocSnapshot::capture();
-        let delta = snap.delta_from(&snap);
-        assert_eq!(delta.allocations, 0);
-        assert_eq!(delta.bytes_allocated, 0);
-        assert_eq!(delta.net_growth(), 0);
     }
 }
